requirement.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
google-api-core==2.0.1
google-auth==2.1.0
google-cloud-vision==2.4.2
googleapis-common-protos==1.53.0
greenlet==1.1.1
grpcio==1.40.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
opencv-python==4.5.3.56
packaging==21.0
Pillow==8.3.2
proto-plus==1.19.0
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.25
urllib3==1.26.6
Werkzeug==2.0.1
